requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/ML Libraries
torch==2.1.0
torchvision==0.16.0
tensorflow==2.14.0
opencv-python==4.8.1.78
deepface==0.0.79
mediapipe==0.10.7
ultralytics==8.0.200  # YOLOv8/v12
efficientnet-pytorch==0.7.1

# NLP and APIs
openai==1.6.1
keybert==0.8.3
transformers==4.35.2
google-api-python-client==2.111.0
youtube-data-api==0.0.21

# Data Processing
numpy==1.24.3
pandas==2.1.4
pillow==10.1.0
scikit-learn==1.3.2
scipy==1.11.4

# Database and Storage
pymongo==4.6.1
redis==5.0.1
boto3==1.34.7

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.3
requests==2.31.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging and Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Development Tools
black==23.12.1
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0